Restart tt* services after edits

+++ src/middlewared/middlewared/etc_files/ttys_config.py
@@ -1,6 +1,20 @@
 import os
+import platform
 import signal
+import subprocess
+
+IS_LINUX = platform.system().lower() == 'linux'
 
 
 def render(service, middleware):
-    os.kill(1, signal.SIGHUP)
+    if IS_LINUX:
+        for command in [
+            ['systemctl', 'restart', 'getty@tty1.service'],
+            ['systemctl', 'restart', 'serial-getty@ttyS0.service'],
+        ]:
+            proc = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            stdout, stderr = proc.communicate()
+            if proc.returncode:
+                middleware.logger.debug('Failed to execute "%s": %s', ' '.join(command), stderr.decode())
+    else:
+        os.kill(1, signal.SIGHUP)
